+++ tools/hs-relmon/src/lib.rs
@@ -7,6 +7,7 @@ pub mod gitlab;
 pub mod list_issues;
 pub mod manifest;
 pub mod prune_tags;
+pub mod review;
 
 pub use sandogasa_repology as repology;
 pub use sandogasa_rpmvercmp as rpmvercmp;